login and register forms

+++ src/main/java/com/betaplan/anjeza/store/controller/ProductController.java
@@ -3,6 +3,7 @@ package com.betaplan.anjeza.store.controller;
 import com.betaplan.anjeza.store.model.Product;
 import com.betaplan.anjeza.store.service.CategoryService;
 import com.betaplan.anjeza.store.service.ProductService;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -28,20 +29,29 @@ public class ProductController {
 
     @GetMapping("/all-products")
     public String showAllProducts(Model model) {
+
         List<Product> products = service.getAllProducts();
         model.addAttribute("products", products);
         return "index";
     }
 
     @GetMapping("/product/new")
-    public String addNewProduct(Model model) {
+    public String addNewProduct(Model model, HttpSession session) {
+        // Check if the user is logged in (check if session contains loggedInUser)
+        if (session.getAttribute("userId") == null) {
+            return "redirect:/login-register";  // Redirect to login page if not logged in
+        }
+
         model.addAttribute("product", new Product());
         model.addAttribute("categories", categoryService.findAll());
         return "product";
     }
 
     @PostMapping("/product/create")
-    public String createProduct(@Valid @ModelAttribute Product product, BindingResult result) {
+    public String createProduct(@Valid @ModelAttribute Product product, BindingResult result, HttpSession session) {
+        if (session.getAttribute("userId") == null) {
+            return "redirect:/login";  // Redirect to login page if not logged in
+        }
         if (result.hasErrors()) {
             return "product";
         } else {
@@ -51,14 +61,24 @@ public class ProductController {
     }
 
     @GetMapping("/product/edit/{id}")
-    public String editExistingProduct(@PathVariable Integer id, Model model) {
+    public String editExistingProduct(@PathVariable Integer id, Model model, HttpSession session) {
+
+        if (session.getAttribute("userId") == null) {
+            return "redirect:/login";  // Redirect to login page if not logged in
+        }
+
         Product product = service.getProductById(id);
         model.addAttribute("product", product);
         return "updateProduct";
     }
 
     @PutMapping("/product/update/{id}")
-    public String updateProduct(@PathVariable Integer id, @Valid @ModelAttribute Product product, BindingResult result) {
+    public String updateProduct(@PathVariable Integer id, @Valid @ModelAttribute Product product, BindingResult result, HttpSession session) {
+
+        if (session.getAttribute("userId") == null) {
+            return "redirect:/login";  // Redirect to login page if not logged in
+        }
+
         if (result.hasErrors()) {
             return "updateProduct";
         } else {
@@ -69,7 +89,12 @@ public class ProductController {
 
     //    @DeleteMapping("/product/delete/{id}")
     @GetMapping("/product/delete/{id}")
-    public String deleteProduct(@PathVariable Integer id) {
+    public String deleteProduct(@PathVariable Integer id, HttpSession session) {
+
+        if (session.getAttribute("userId") == null) {
+            return "redirect:/login";  // Redirect to login page if not logged in
+        }
+
         service.deleteProductById(id);
         return "redirect:/all-products";
     }
